private/requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
flask_cors==3.0.10
recipe_scrapers
openai==1.12.0
python_dotenv==1.0.0
beautifulsoup4==4.12.2
requests==2.31.0